requirements.txt
openai>=1.30.0
feedparser>=6.0.10
requests>=2.31.0
Jinja2>=3.1.4
PyYAML>=6.0.1
tenacity>=8.2.3